+++ src/iosanita/contenttypes/interfaces/struttura.py
@@ -75,7 +75,7 @@ class IStruttura(model.Schema, IIosanitaContenttypes):
         ),
         description=_(
             "personale_struttura_struttura_help",
-            default="Personale della struttura.",
+            default='Utilizza questo campo per definire manualmente una lista delle persone da mostrare come personale della struttura. Se le impostazioni del sito lo prevedono, a questo elenco verranno aggiunte le Persone che hanno referenziato questa struttura nei campi "dove opera" o "dove riceve"',
         ),
         required=False,
         default=[],
@@ -127,18 +127,18 @@ class IStruttura(model.Schema, IIosanitaContenttypes):
         vocabulary="plone.app.vocabularies.Catalog",
         pattern_options={"selectableTypes": ["Persona"]},
     )
-    form.widget(
-        "personale_struttura",
-        RelatedItemsFieldWidget,
-        vocabulary="plone.app.vocabularies.Catalog",
-        pattern_options={"selectableTypes": ["Persona"]},
-    )
 
     form.widget(
         "uo_correlata",
         RelatedItemsFieldWidget,
         vocabulary="plone.app.vocabularies.Catalog",
         pattern_options={"selectableTypes": ["UnitaOrganizzativa"]},
+    )
+    form.widget(
+        "personale_struttura",
+        RelatedItemsFieldWidget,
+        vocabulary="plone.app.vocabularies.Catalog",
+        pattern_options={"selectableTypes": ["Persona"]},
     )
 
     # fieldsets